Ignore changed  .gemspec from vendor/cache folder
When we vendor a bundler project we often end up with an updated
.gemspec files in the vendor folder but we don't care about these when
creating the PR branch or message as this incorrectly identfies the
dependency as a library and attempts to generate a updated requirement
but this sometimes doesn't change when only bumping the version in the
lockfile.

This is a quick fix to keep the existing logic but we should refactor
this to get it in line with the existing logic to determine if a project
is a library as this varies between package manager.